Remove redundant function from PriorityQueueHandle
The context in PriorityQueue::invalidateHandleOf makes it clear that
a default handle is invalid.
It is not needed even outside of it because clients can instantiate only
default handles. They should not even care if it is valid or not, since they
have no other choice. (Althoug it is not hard to guess it is invalid)