Extract a private function in PriorityQueue
Extracted a function from buildHeap(). It computes the
position of the last non-leaf element in the heap to start
sifting from there.
It also uses traits to preserve the size and sign of
the result index (which must be signed).
The function is inlined in the class definition because it uses auto
to deduce the return type. (The other option was to
make the asSigned lambda in it a separate function but
it makes no sense outside of it)