Fix a bug in Hash::remove
The problem was that if when shrinking the table
an exception was thrown, the extracted item was not
inserted back and thus the function was not exception safe.
This is now fixed altough the code is not as clear as before.
This is the most clear way I came up with in order to make the
function offer the strong exception guarantee.