Move the EdgeList out of Vertex
Replaced the EdgeList with iterator to a collection which will
contain all the EdgeLists of a graph's vertices.
I did this to make Vertex objects small enough to be contained in an array by value
and not-so-expensive to copy.
This still leaves the knowledge of edges in a Vertex but is in the name of efficiency.
If there is no iterator the edges will have to be searched in a map or
something similar each time an algorithm for searching a graph needs them.
This would be a massive bottleneck.