Loosen some type requirements of PriorityQueue's template arguments
The functors (functions) need not be nothrow default constructible etc.
in order the algorithm to work. This must be true only if
clients want PriorityQueue to have noexcept default ctor, move ctor etc.
The arguments operator()'s on the other side are required to be noexcept
in order the algorithm to be safe.