Make GraphCreator take the ID by reference
The ID is passed all the way from the command through
the factory, the creator and then the Graph's ctor.
When the factory has an Rvalue overload, things could change
but for now there is no need in moving the ID so many times.
Moreover, it may not be copied since the ID could be rejected.

I left the constructor take the string by value because
it is passed only once - from GraphRegistrator's ctor
(which also takes by value because literals are passed to it)
to the GraphCreator's ctor. This will be faster than making
a temporary for Registrator and then copying it again.